Split selection edit controller actions
